% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion.R
\name{degreeStratifiedMonteCarlo}
\alias{degreeStratifiedMonteCarlo}
\title{Median-based degree-stratified Monte Carlo test}
\usage{
degreeStratifiedMonteCarlo(
  network,
  sample,
  method = "sturges",
  diffusion.scores,
  permutations = 1000,
  plottitle = ""
)
}
\arguments{
\item{network}{igraph object}

\item{sample}{Character vector of gene IDs to test}

\item{method}{Character; discretization method}

\item{diffusion.scores}{Named numeric vector of diffusion scores}

\item{permutations}{Integer; number of replicates}

\item{plottitle}{Character; title for density plots}
}
\value{
List with Permutations.Median, Test.Median, Test.Percentile
}
\description{
Median-based degree-stratified Monte Carlo test
}
