% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion.R
\name{degreeStratifiedMonteCarloMean}
\alias{degreeStratifiedMonteCarloMean}
\title{Mean-based degree-stratified Monte Carlo test}
\usage{
degreeStratifiedMonteCarloMean(
  network,
  sample,
  method = "sturges",
  diffusion.scores,
  permutations = 1000
)
}
\arguments{
\item{network}{igraph object}

\item{sample}{Character vector of gene IDs to test}

\item{method}{Character; discretization method}

\item{diffusion.scores}{Named numeric vector of diffusion scores}

\item{permutations}{Integer; number of replicates}
}
\value{
List with Permutations.Mean, Test.Mean, Test.Percentile
}
\description{
Mean-based degree-stratified Monte Carlo test
}
