% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion.R
\name{degreeStratifiedSample}
\alias{degreeStratifiedSample}
\title{Degree-stratified sampling for control selection}
\usage{
degreeStratifiedSample(
  network,
  sample,
  control,
  sample_size,
  method = "sturges"
)
}
\arguments{
\item{network}{igraph object}

\item{sample}{Character vector of focal IDs}

\item{control}{Character vector of IDs to sample from}

\item{sample_size}{Integer; number to sample}

\item{method}{Character; discretization method}
}
\value{
Data.frame of sampled gene metadata
}
\description{
Degree-stratified sampling for control selection
}
