% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart_helpers.R
\name{DownloadGenesInGO}
\alias{DownloadGenesInGO}
\title{Download Genes Annotated to a GO Term from Ensembl (Human)}
\usage{
DownloadGenesInGO(
  request,
  text = FALSE,
  children = TRUE,
  attributes = c("ensembl_gene_id", "entrezgene_id", "external_gene_name")
)
}
\arguments{
\item{request}{Character; GO term ID}

\item{text}{Logical; if TRUE, write CSV to working dir}

\item{children}{Logical; if TRUE, include child terms}

\item{attributes}{Character vector; Ensembl attributes}
}
\value{
data.frame of annotated genes
}
\description{
Download Genes Annotated to a GO Term from Ensembl (Human)
}
