% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_analysis.R
\name{findAutophagyDoublePapers}
\alias{findAutophagyDoublePapers}
\title{Get counts for a term and term+autophagy overlap}
\usage{
findAutophagyDoublePapers(query, statsOnly = TRUE)
}
\arguments{
\item{query}{Character; PubMed query}

\item{statsOnly}{Logical; if TRUE, return counts only}
}
\value{
Numeric vector: c(totalCount, autophagyCount) or full PubMed objects
}
\description{
Get counts for a term and term+autophagy overlap
}
