% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_analysis.R
\name{findDoubleQueryPapers}
\alias{findDoubleQueryPapers}
\title{Perform a PubMed AND query of two terms}
\usage{
findDoubleQueryPapers(query1, query2)
}
\arguments{
\item{query1, query2}{Character; individual PubMed queries}
}
\value{
A PubMed ID object from easyPubMed::get_pubmed_ids()
}
\description{
Perform a PubMed AND query of two terms
}
