% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart_helpers.R
\name{getGenesFromList}
\alias{getGenesFromList}
\title{Map a list of identifiers to gene metadata}
\usage{
getGenesFromList(
  input = NULL,
  filters = "ensembl_gene_id",
  attributes = c("ensembl_gene_id", "entrezgene_id", "external_gene_name")
)
}
\arguments{
\item{input}{Vector of identifiers.}

\item{filters}{BioMart filter name.}

\item{attributes}{Character vector of BioMart attributes to return.}
}
\value{
A data frame of mapped identifiers.
}
\description{
Map a list of identifiers to gene metadata
}
