% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion.R
\name{getGONetworkNodes}
\alias{getGONetworkNodes}
\title{Get GO-annotated network nodes}
\usage{
getGONetworkNodes(go, network, attr = "entrezgene_id")
}
\arguments{
\item{go}{Character; GO term ID}

\item{network}{igraph object}

\item{attr}{Character; attribute in GO result to match (e.g. "entrezgene_id")}
}
\value{
Character vector of vertex names in network
}
\description{
Get GO-annotated network nodes
}
