% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_analysis.R
\name{getTermPapers}
\alias{getTermPapers}
\title{Build a table of autophagy stats for multiple terms}
\usage{
getTermPapers(term, table)
}
\arguments{
\item{term}{Character; a single query term}

\item{table}{List of terms}
}
\value{
Matrix of counts for each term
}
\description{
Build a table of autophagy stats for multiple terms
}
