% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion.R
\name{getTopList}
\alias{getTopList}
\title{Retrieve top-scoring genes with metadata}
\usage{
getTopList(number = 200, list)
}
\arguments{
\item{number}{Integer; number of top entries}

\item{list}{Named numeric vector of scores}
}
\value{
Data.frame of gene metadata plus diffusion_score
}
\description{
Retrieve top-scoring genes with metadata
}
