% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion.R
\name{logVariableDiscretisation}
\alias{logVariableDiscretisation}
\title{Discretise a numeric variable after log-transform}
\usage{
logVariableDiscretisation(variable, names, method = "sturges")
}
\arguments{
\item{variable}{Numeric vector}

\item{names}{Character vector of IDs (same length as variable)}

\item{method}{Character; discretization method}
}
\value{
Data.frame with ID, Variable, LogVariable, VarDiscret, LogVarDiscret
}
\description{
Discretise a numeric variable after log-transform
}
