% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion.R
\name{networkDiscretisation}
\alias{networkDiscretisation}
\title{Discretise network by log-degree}
\usage{
networkDiscretisation(network, method = "sturges")
}
\arguments{
\item{network}{igraph object}

\item{method}{Character; discretization method (e.g. "sturges")}
}
\value{
Data.frame with Gene.ID, Log.Degree, Category
}
\description{
Discretise network by log-degree
}
