% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion.R
\name{plotPRC}
\alias{plotPRC}
\title{Plot Precision–Recall curve}
\usage{
plotPRC(scores, correct)
}
\arguments{
\item{scores}{Numeric vector of scores}

\item{correct}{Logical vector of true/false labels}
}
\value{
Invisible; side-effect plot plus printed AUC (PR)
}
\description{
Plot Precision–Recall curve
}
