% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion.R
\name{plotROC}
\alias{plotROC}
\title{Plot ROC curve from scores}
\usage{
plotROC(scores, correct)
}
\arguments{
\item{scores}{Numeric vector of scores}

\item{correct}{Logical vector of true/false labels}
}
\value{
Invisible; side-effect plot plus printed AUC
}
\description{
Plot ROC curve from scores
}
