% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion.R
\name{stratifiedSample}
\alias{stratifiedSample}
\title{Stratified sampling by a continuous control variable}
\usage{
stratifiedSample(
  allNames,
  sample,
  controlVariable,
  controlSet,
  sample_size,
  method = "sturges"
)
}
\arguments{
\item{allNames}{Character vector of all IDs}

\item{sample}{Character vector of IDs to sample}

\item{controlVariable}{Numeric vector (same length as allNames)}

\item{controlSet}{Character vector; pool to sample from}

\item{sample_size}{Integer; size of sample}

\item{method}{Character; discretization method}
}
\value{
Data.frame of sampled gene metadata
}
\description{
Stratified sampling by a continuous control variable
}
