% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion.R
\name{targetBoxPlot}
\alias{targetBoxPlot}
\title{Boxplot of diffusion scores for target vs non-target}
\usage{
targetBoxPlot(
  diffusionScores,
  network,
  positives,
  target = NULL,
  method = NULL,
  positiveCategory = NULL,
  category2 = NULL,
  category2name = "Category 2",
  returnData = FALSE
)
}
\arguments{
\item{diffusionScores}{Named numeric vector of diffusion scores}

\item{network}{igraph object}

\item{positives}{Character vector of “positive” gene IDs}

\item{target}{Character vector of genes to include (defaults to all)}

\item{method}{Character; diffusion method label}

\item{positiveCategory}{Character; label for positives}

\item{category2}{Character vector of a second category (optional)}

\item{category2name}{Character; label for second category}

\item{returnData}{Logical; if TRUE, return the data.frame instead of plotting}
}
\value{
ggplot2 object or data.frame if returnData = TRUE
}
\description{
Boxplot of diffusion scores for target vs non-target
}
