% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion.R
\name{testGO}
\alias{testGO}
\title{Degree-stratified test for a GO term}
\usage{
testGO(go, network, attr = "entrezgene_id", diffusion.scores, perm = 1000)
}
\arguments{
\item{go}{Character; GO term ID}

\item{network}{igraph object}

\item{attr}{Character; attribute in GO result to match (e.g. "entrezgene_id")}

\item{diffusion.scores}{Named numeric vector}

\item{perm}{Integer; number of permutations}
}
\value{
Result of degreeStratifiedMonteCarlo()
}
\description{
Degree-stratified test for a GO term
}
