% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion.R
\name{topScoresSubnetwork}
\alias{topScoresSubnetwork}
\title{Plot the top-scoring subnetwork}
\usage{
topScoresSubnetwork(diffusionScores, known_positive, number, network)
}
\arguments{
\item{diffusionScores}{Named numeric vector of diffusion scores}

\item{known_positive}{Character vector of known positives}

\item{number}{Integer; how many top nodes to include}

\item{network}{igraph object}
}
\value{
Invisible; side-effect plot
}
\description{
Plot the top-scoring subnetwork
}
