% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion.R
\name{topXpercentIntersection}
\alias{topXpercentIntersection}
\title{Intersection of top X percent lists}
\usage{
topXpercentIntersection(network, list1, list2, percent = 0.05)
}
\arguments{
\item{network}{An igraph object.}

\item{list1}{Named numeric vector}

\item{list2}{Named numeric vector}

\item{percent}{Numeric; fraction of network to take}
}
\value{
Character vector of intersecting external_gene_name
}
\description{
Intersection of top X percent lists
}
